Moved killAndReopenApp to Page class
